2011-02-01 Roman - SYS: Build with all warning show is enabled for RPM-packages. - SYS: The default start OpenSCADA scripts "openscada_start" "openscada_demo" and "openscada_demo_local" is adapted     for use simple shell and don't use tool install, on busybox. - FIX:SYS: Code clear for all source code:     - remove not used variables;     - add returns for all functions with return;     - init variables which possible used undefined;     - not full enumerated variables processing into switches;     - signed and unsigned comparision. - FIX:Archive: The prototype of functions TArchiveS::messPut() and TArchiveS::messGet() is changed for fix signed     message level on ARM CPUs. - FIX:SYS: The prototype of functions TMess::put(), TMess::get(), TCntrNode::grpSize(), TCntrNode::grpId(), TCntrNode::grpAt()     is changed for fix unsigned char on ARM CPUs. - FIX:DAQ: DAQ-redundant logic some fixed. - FIX:Mess: Default language select logic is some fixed. - FIX:Transport: All transports loading is fixed for missed some transport modules and DB records for its allow. - FIX:Value: The prototype of functions TValue::setReqFlg(), TValue::setResB1() and TValue::setResB2() is fixed for no return. - FIX:XML: Expanded variable size for flags of functions XMLNode::save(). - FIX:Archive.FSArch: Result from fgetc() is writed to integer without char and fix "EOF" detection.     File detect is fixed into ModVArchEl::fileAdd(). - FIX:DB.MySQL: case TFld::String missing is fixed into switch of function MTable::fieldFix(). - FIX:DAQ.JavaLikeCalc: Rezult value type set is fixed for unary operations.     The function string::parsePath() of user API is fixed for offset use.     Possible disabling functions for return operations. - Protocol.SelfSystem: Enter value limits is set to main configuration parameters of the module. - FIX:UI.Vision: For get painted text width used fontMetrics().width() without fontMetrics().boundingRect(lab_tm.c_str()).width() for     prevent some fluctuations.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1254 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/OPC_UA/mod_prt.cpp
@@ -1068,7 +1068,7 @@ void TProt::oR( string &buf, double val, char sz )
 
 void TProt::oS( string &buf, const string &val )
 {
-    oN(buf,(val.size()?val.size():-1),4);
+    oN(buf,(val.size()?(int)val.size():-1),4);
     buf.append(val);
 }
